refactor: refactor ManufacturingOrderIconButtonList
